Completed 9 tests for the Student Model across the CRUD actions.
The Create action has one test that passes and two that fail. The Read, Update, and Destroy actions each have one test that passes and one that fails. I suspect that at least a couple of my tests are not particularly useful or are not operating in the way I expect based solely on the syntax and my general unfamiliarity with Ruby.